refactoring and make the API return error messages
