+++ Transaction/TransactionHelper.h
@@ -9,12 +9,12 @@
 #include "../BlockHeader.h"
 #include "../Scripts/DeactivateCertificateScript.h"
 #include "../CertStore/Cert.h"
+#include "TransactionError.h"
 
 class TransactionHelper {
-private:
+public:
     static bool verifyNonce(std::vector<unsigned char> inAddress, uint32_t nonce);
     static uint32_t getNonce(std::vector<unsigned char> inAddress);
-public:
     static std::vector<unsigned char> getDeactivateCertificateScriptId(DeactivateCertificateScript deactivateCertificateScript);
     static std::vector<unsigned char> getTxId(Transaction* tx);
     static std::vector<unsigned char> getTxHash(Transaction* tx);
@@ -22,11 +22,6 @@ public:
     static std::vector<unsigned char> getPassportHash(Transaction* tx, X509* x509);
     static bool isVote(Transaction* tx);
     static bool isRegisterPassport(Transaction* tx);
-    static bool verifyNetworkTx(TransactionForNetwork* tx);
-    static bool verifyTx(Transaction* tx, uint8_t isInHeader,  BlockHeader* header);
-    static bool verifyRegisterPassportTx(Transaction* tx, uint32_t blockHeight, Cert* cert);
-    static bool applyTransaction(Transaction* tx, BlockHeader* blockHeader);
-    static bool undoTransaction(Transaction* tx, BlockHeader* blockHeader);
     static UAmount calculateMinimumFee(Transaction* transaction, BlockHeader* header);
     static UAmount calculateMinimumFee(size_t txSize, BlockHeader* header);
 };
